Refactor: fix loop
