+++ lib/screen/interactive_postcard/travel_info/travel_info_state.dart
@@ -1,0 +1,7 @@
+import 'package:autonomy_flutter/model/travel_infor.dart';
+
+class TravelInfoState {
+  final List<TravelInfo>? listTravelInfo;
+
+  TravelInfoState({this.listTravelInfo});
+}
